+++ traits/TimeZoneTrait.php
@@ -8,26 +8,38 @@ use \DateTimeZone;
 # 날짜 관련 데이터베이스 저장 및 뷰
 trait TimeZoneTrait
 {
-  public function nowInTZ(string $utcgmttime): string
-  {
-    return (new DateTimez("now", $utcgmttime))->format('Y-m-d H:i:s P');
-  }
+	public function nowInTZ(
+		string $utcgmttime
+	): string
+	{
+		return (new DateTimez("now", $utcgmttime))
+		->format('Y-m-d H:i:s P');
+	}
 
-  public function dataInTZ(string $data, string $utcgmttime, string $format="Y-m-d H:i:s P"): string
-  {
-    return (new DateTimez($data, $utcgmttime))->format($format);
-  }
+	public function toTZFormat(
+		string $datetimeptz,  
+		string $from, 
+		string $to, 
+		array|string $timezone_formats
+	) : ?string
+	{
+		if(!$datetimeptz){
+			return null;
+		}
 
-  public function toTZFormat(string $datetimeptz,  string $utcgmttime, string $convert_utcgmttime, array $timezone_formats) : ?string
-  {
-    if(!$datetimeptz){
-      return null;
-    }
+		$dataTimeZ = (new DateTimez($datetimeptz, $from));
+		$dataTimeZ->setTimezone(new DateTimeZone($to));
 
-    $dataTimeZ = (new DateTimez($datetimeptz, $utcgmttime));
-    $dataTimeZ->setTimezone(new DateTimeZone($convert_utcgmttime));
-    return $dataTimeZ->format(
-      ((new ArrayHelper( $timezone_formats ))->find("timezone",$convert_utcgmttime)->value)['format']
-    );
-  }
+		$format = '';
+		if(is_array($timezone_formats)){
+			$format = ((new ArrayHelper( $timezone_formats ))->find("timezone",$to)->value)['format'];
+		}else if(is_string($timezone_formats)){
+			$format = $timezone_formats;
+		}
+
+		if(!$format)
+			return null;
+
+		return $dataTimeZ->format($format);
+	}
 }