+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 ## Banana[3.5.7]
-### - 2025-07-11
-- TimeZoneTrait dataInTZ method new add
+### - 2025-07-14
+- TimeZoneTrait toTZFormat 업데이트 timezone_formats 을 array | string 두가지 타입으로 가능하도록 업데이트
 
 ### - 2025-07-08
 - TextUtil v1.1->1.2 각 파트별 성능 개선 (mbstring 필수 )